api/requirements.txt
# For the server
gunicorn==20.0.4

# For the application
flask==1.1.2
pymongo==3.11.4
dnspython==2.1.0
paho-mqtt==1.5.1